final Scripts/requirements.txt
altgraph==0.17.4
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
lxml==5.1.0
numpy==1.26.3
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
tzdata==2023.4
urllib3==2.1.0
wsproto==1.2.0
altgraph==0.17.4
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
lxml==5.1.0
numpy==1.26.3
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
tzdata==2023.4
urllib3==2.1.0
wsproto==1.2.0
